feat: :recycle: Refactorando pastas e arquivos para melhoria dos componentes e iniciando menu lateral.
