+++ src/App.js
@@ -1,12 +1,14 @@
 import Cabecalho from './components/Cabecalho';
+import MenuNavegacao from 'components/Menu_Navegacao';
+import MenuLateral from 'components/Menu_Lateral';
 import './App.css';
-import MenuLateral from 'components/MenuNavegacao';
 
 function App() {
   return (
     <>
       <Cabecalho />
-      <MenuLateral />
+      <MenuNavegacao />
+      <MenuLateral/>
     </>
   );
 }
