--- src/components/InputPesquisa/index.jsx
@@ -1,10 +1,0 @@
-import styles from './InputPesquisa.module.css';
-
-const InputPesquisa = ({text = 'text', placText}) => {
-    return(
-        <input type={text} placeholder={placText} className={styles.pesquisa}/>
-    )
-
-}
-
-export default InputPesquisa;
