+++ src/components/Input_Pesquisa/InputPesquisa.module.css
@@ -2,8 +2,26 @@
     display: none;
 }
 
-@media (min-width: 768px) {
+.input {
+    display: block;
+    height: 54px;
+    border: none;
+    background-color: rgba(255, 255, 255, 0.2);
+    color: var(--branco);
+    border-radius: 8px;
+    outline: none;
+    padding: 0 1em;
+}
 
+.input::placeholder {
+    color: rgba(255, 255, 255, 0.5);
+    line-height: 24px;
+    font-size: 16px;
+    font-family: var(--font);
+    font-weight: 400;
+}
+
+@media (min-width: 768px) {
     .pesquisa {
         display: block;
         width: 431px;
@@ -16,16 +34,13 @@
         padding: 0 1em;
     }
 
-    .pesquisa::placeholder,
-    .pesquisa {
+    .pesquisa::placeholder {
         color: rgba(255, 255, 255, 0.5);
         line-height: 24px;
         font-size: 16px;
         font-family: var(--font);
         font-weight: 400;
     }
-
-
 }
 
 @media (min-width: 1024px) {
