+++ src/components/Menu_Lateral/MenuLateral.module.css
@@ -1,0 +1,7 @@
+.menu_lateral{
+    padding: 1em;
+} 
+form {
+    display: flex;
+    flex-direction: column;
+}
