+++ src/components/Menu_Lateral/index.jsx
@@ -1,0 +1,16 @@
+import Titulo from 'components/Titulo';
+import styles from './MenuLateral.module.css';
+import InputPesquisa from 'components/Input_Pesquisa';
+
+const MenuLateral = () => {
+    return(
+        <aside className={styles.menu_lateral}>
+            <form>
+                <Titulo>Seu Projeto</Titulo>
+                <InputPesquisa placText='Nome do seu projeto' text='text' stilos='input'/>
+            </form>
+        </aside>
+    )
+};
+
+export default MenuLateral
