+++ src/components/Menu_Navegacao/MenuNavegacao.module.css
@@ -8,14 +8,7 @@
         display: flex;
         flex-direction: column;
         
-    }
-    
-    .menu_navegacao h2 {
-        padding: 1em 0;
-        color: var(--branco);
-        font-family: var(--font);
-    }
-    
+    }    
     .menu_navegacao ul {
         display: flex;
         flex-direction: column;
