+++ src/components/Menu_Navegacao/index.jsx
@@ -1,12 +1,13 @@
-import MenuLista from "./MenuLista";
+import MenuLista from "./Menu_Lista";
 import styles from "./MenuNavegacao.module.css";
 import iconCode from "./IconCode.png";
 import iconComunidade from "./iconComunidade.png";
+import Titulo from "components/Titulo";
 
 const MenuNavegacao = () => {
     return(
         <aside className={styles.menu_navegacao}>
-            <h2>Menu</h2>
+            <Titulo>Menu</Titulo>
                 <ul>
                     <MenuLista children='Editor de código' imagem={iconCode} link={"#"} texto={'icone code'}/>
                     <MenuLista children='Comunidade' imagem={iconComunidade} link={"#"} texto={'icone code'}/>
